fix: address consolidated model review
