+++ scripts/hf-skippy-convert-job.py
@@ -8,6 +8,7 @@
 from __future__ import annotations
 
 import argparse
+import json
 import os
 import shutil
 import subprocess
@@ -134,6 +135,7 @@ def convert(args: argparse.Namespace, root: Path) -> Path:
     if status.exists():
         shutil.copy2(status, artifact_dir / "skippy-convert-status.json")
     write_beta_card(artifact_dir, args.source_repo, args.mesh_revision)
+    validate_converted_artifact(artifact_dir)
     return artifact_dir
 
 
@@ -149,16 +151,42 @@ def upload(args: argparse.Namespace, artifact_dir: Path) -> None:
     )
 
 
+def validate_converted_artifact(artifact_dir: Path) -> None:
+    required_files = ("README.md", "skippy-convert-manifest.json")
+    missing = [name for name in required_files if not (artifact_dir / name).is_file()]
+    if not artifact_dir.is_dir() or missing:
+        details = f"; missing {', '.join(missing)}" if missing else ""
+        raise FileNotFoundError(
+            f"complete converted artifact not found: {artifact_dir}{details}"
+        )
+
+    manifest_path = artifact_dir / "skippy-convert-manifest.json"
+    manifest = json.loads(manifest_path.read_text(encoding="utf-8"))
+    expected_splits = manifest.get("expected_splits")
+    basename = manifest.get("output_basename")
+    if not isinstance(expected_splits, int) or expected_splits < 1:
+        raise ValueError(f"invalid expected_splits in {manifest_path}")
+    if not isinstance(basename, str) or not basename:
+        raise ValueError(f"invalid output_basename in {manifest_path}")
+
+    if expected_splits == 1:
+        expected_names = [f"{basename}.gguf"]
+    else:
+        expected_names = [
+            f"{basename}-{index:05}-of-{expected_splits:05}.gguf"
+            for index in range(1, expected_splits + 1)
+        ]
+    missing_shards = [name for name in expected_names if not (artifact_dir / name).is_file()]
+    if missing_shards:
+        raise FileNotFoundError(
+            f"converted artifact is incomplete: missing {', '.join(missing_shards)}"
+        )
+
+
 def converted_artifact_dir(args: argparse.Namespace) -> Path:
     artifact_dir = Path(args.work_dir) / "target" / args.target_prefix
     if args.upload_only:
-        required_files = ("README.md", "skippy-convert-manifest.json")
-        missing = [name for name in required_files if not (artifact_dir / name).is_file()]
-        if not artifact_dir.is_dir() or missing or not any(artifact_dir.glob("*.gguf")):
-            details = f"; missing {', '.join(missing)}" if missing else ""
-            raise FileNotFoundError(
-                f"complete converted artifact not found: {artifact_dir}{details}"
-            )
+        validate_converted_artifact(artifact_dir)
     return artifact_dir
